+++ api/baeuja/build.gradle
@@ -5,7 +5,7 @@ plugins {
 	id 'org.asciidoctor.jvm.convert' version '3.3.2'
 }
 
-group = 'we-are-crack'
+group = 'xyz.baeuja'
 version = '0.0.1-SNAPSHOT'
 
 java {
