--- api/baeuja/src/main/java/we_are_crack/baeuja/global/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package we_are_crack.baeuja.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
