--- api/baeuja/src/main/java/we_are_crack/baeuja/global/exception/DuplicateNicknameException.java
@@ -1,22 +1,0 @@
-package we_are_crack.baeuja.global.exception;
-
-public class DuplicateNicknameException extends RuntimeException {
-
-    public static final String CODE = "DUPLICATE_NICKNAME";
-
-    public DuplicateNicknameException() {
-        super("이미 사용 중인 닉네임입니다.");
-    }
-
-    public DuplicateNicknameException(String message) {
-        super(message);
-    }
-
-    public DuplicateNicknameException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DuplicateNicknameException(Throwable cause) {
-        super(cause);
-    }
-}
