--- api/baeuja/src/main/java/we_are_crack/baeuja/global/exception/InvalidNicknameException.java
@@ -1,22 +1,0 @@
-package we_are_crack.baeuja.global.exception;
-
-public class InvalidNicknameException extends RuntimeException {
-
-    public static final String CODE = "INVALID_NICKNAME";
-
-    public InvalidNicknameException() {
-        super("유효하지 않은 닉네임입니다.");
-    }
-
-    public InvalidNicknameException(String message) {
-        super(message);
-    }
-
-    public InvalidNicknameException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InvalidNicknameException(Throwable cause) {
-        super(cause);
-    }
-}
