--- api/baeuja/src/main/java/we_are_crack/baeuja/global/exception/UnexpectedException.java
@@ -1,22 +1,0 @@
-package we_are_crack.baeuja.global.exception;
-
-public class UnexpectedException extends RuntimeException {
-
-    public static final String CODE = "INTERNAL_ERROR";
-
-    public UnexpectedException() {
-        super("서버 내부 오류가 발생했습니다.");
-    }
-
-    public UnexpectedException(String message) {
-        super(message);
-    }
-
-    public UnexpectedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UnexpectedException(Throwable cause) {
-        super(cause);
-    }
-}
