--- api/baeuja/src/main/java/we_are_crack/baeuja/user/domain/LoginType.java
@@ -1,5 +1,0 @@
-package we_are_crack.baeuja.user.domain;
-
-public enum LoginType {
-    GUEST, GOOGLE
-}
