+++ api/baeuja/src/main/java/xyz/baeuja/api/BaeujaApplication.java
@@ -1,0 +1,13 @@
+package xyz.baeuja.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BaeujaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BaeujaApplication.class, args);
+	}
+
+}
