+++ api/baeuja/src/main/java/xyz/baeuja/api/global/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package xyz.baeuja.api.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
