+++ api/baeuja/src/main/java/xyz/baeuja/api/global/exception/DuplicateEmailException.java
@@ -1,0 +1,22 @@
+package xyz.baeuja.api.global.exception;
+
+public class DuplicateEmailException extends RuntimeException {
+
+    public static final String CODE = "DUPLICATE_EMAIL";
+
+    public DuplicateEmailException() {
+        super("이미 사용 중인 이메일입니다.");
+    }
+
+    public DuplicateEmailException(String message) {
+        super(message);
+    }
+
+    public DuplicateEmailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicateEmailException(Throwable cause) {
+        super(cause);
+    }
+}
