+++ api/baeuja/src/main/java/xyz/baeuja/api/user/domain/LoginType.java
@@ -1,0 +1,5 @@
+package xyz.baeuja.api.user.domain;
+
+public enum LoginType {
+    GUEST, GOOGLE
+}
