+++ api/baeuja/src/main/java/xyz/baeuja/api/user/domain/User.java
@@ -1,0 +1,50 @@
+package xyz.baeuja.api.user.domain;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import xyz.baeuja.api.global.entity.BaseEntity;
+
+@Entity
+@Table(name = "users")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id")
+    private Long id;
+
+    @Column(length = 100, unique = true)
+    private String email;
+
+    @Column(length = 20, nullable = false, unique = true)
+    private String nickname;
+
+    @Column(nullable = false)
+    private String locale;
+
+    @Column(nullable = false)
+    private String timezone;
+
+    @Enumerated(value = EnumType.STRING)
+    @Column(nullable = false)
+    private LoginType loginType;
+
+    public User(String nickname, String locale, String timezone, LoginType loginType) {
+        this.nickname = nickname;
+        this.locale = locale;
+        this.timezone = timezone;
+        this.loginType = loginType;
+    }
+
+    /**
+     * 구글 계정으로 전환
+     */
+    public void convertToGoogleAccount(String email) {
+        this.loginType = LoginType.GOOGLE;
+        this.email = email;
+    }
+}
