--- api/baeuja/src/test/java/we_are_crack/baeuja/BaeujaApplicationTests.java
@@ -1,13 +1,0 @@
-package we_are_crack.baeuja;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BaeujaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
