+++ api/baeuja/src/test/java/xyz/baeuja/api/BaeujaApplicationTests.java
@@ -1,0 +1,13 @@
+package xyz.baeuja.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BaeujaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
